Email alerts & Google Sheets Robust
Added a webhook the pings Zapier any time the map
(a) throws an error in the browser or
(b) loads an empty or null dataset from Google Sheets.

Also, changed the code so that it exclusively uses Google Visualization API tools to unpack the object retrieved from Google Sheets using that API, so that a small change in their data structure doesn't kill the map.