Update plotData() in script.js
Fixed a v. rare issue where uncomplete rows of data could cause plotData() to fail:
Incomplete rows in AllData used to be simply igonred, which could lead the function to try and access AllData[i].duplicates for a point for which it had not been created.
Now those rows are deleted from AllData()